Resubmitting under the MIT License
> From: Bastian Barton <bastian.barton@posteo.de>
> Subject: Re: Relicensing RamanChada under the MIT License
> Date: Sun, 22 Jan 2023 21:00:20 +0000
> Cc: "Lellinger, Dirk" <dirk.lellinger@lbf.fraunhofer.de>, "Нина Желязкова" <jeliazkova.nina@gmail.com>, Georgi Georgiev <georgiev.georgi.st@gmail.com>, James <jamest@elodiz.com>
> To: Luchesar ILIEV <luchesar.iliev@gmail.com>
>
> Happy new year everyone!
>
> I, Bastian Barton, hereby agree to have all my contributions to the H2020 CHARISMA RamanChada project and repository resubmitted under the terms of the MIT License.
>
> My email addresses used for contributions are: Bastian.barton@lbf.Fraunhofer.de

NB: While RFC 5321 mandates the local-part of an email address to be
treated in SMTP as case-sensitive, in practice, most if not all email
backends will process the addresses in a case-insensitive manner, i.e.
will treat <Bastian.barton@lbf.Fraunhofer.de> as being the very same as
<bastian.Barton@lbf.fraunhofer.de> or <Bastian.Barton@lbf.fraunhofer.de>
or <bastian.barton@lbf.fraunhofer.de>. Hence why we assume that all
contributions made under these email identities are covered by the above
agreement even if there may be a case mismatch in the email address.
Note as well that the domain part is always processed in a
case-insensitive manner per RFC 1035.